Removed weird test flag in YahooScraper. Added NullPointerException handler (have to change log message, I copied and pasted it from the IOException above) Rewrote scrapeSingleSummaryData method Added a lot of documentation comments
Wrapping up for the night of 7/10.  Made a tragically stupid mistake and am committing hari-kiri for the night.
As of this commit, the new code works.  All of the DB fields are filled properly.
Next thing on the chopping block is to refactor all the Utility methods into one area ahead of the switch statement.

Refactoring proved harder than anticipated, so I'm keeping the Utility methods inside the case statements.
Cleaned up documentation, removed deprecated code.
Next step is to compare output between the old code and this one and see if they match.  Then finalize the branch.